created new strategy for phpunit tests
