+++ tests/SlmLocaleTest/Strategy/UriPathStrategyTest.php
@@ -348,28 +348,6 @@ class UriPathStrategyTest extends TestCase
         $this->assertEquals($expected, $actual);
     }
 
-    public function testDisableUriPathStrategyPhpunit()
-    {
-        $_SERVER['DISABLE_URIPATHSTRATEGY'] = true;
-
-        $uri     = 'http://username:password@example.com:8080/some/deep/path/some.file?withsomeparam=true';
-        $request = new HttpRequest();
-        $request->setUri($uri);
-
-        $this->event->setLocale('en');
-        $this->event->setRequest($request);
-        $this->event->setResponse(new HttpResponse());
-
-        $this->strategy->found($this->event);
-
-        $statusCode = $this->event->getResponse()->getStatusCode();
-        $header     = $this->event->getResponse()->getHeaders()->get('Location');
-        $expected   = 'Location: http://username:password@example.com:8080/en/some/deep/path/some.file?withsomeparam=true';
-        $this->assertEquals(200, $statusCode);
-
-        $_SERVER['DISABLE_URIPATHSTRATEGY'] = false;
-    }
-
     protected function getPluginManager($console = false)
     {
         $sl = new ServiceManager();
